Upgraded the manipulator of the Section 13 with some new fields management, released as 1.3.4 version

+++ PSE.Model/Output/Interfaces/IBase.cs
@@ -51,7 +51,7 @@ namespace PSE.Model.Output.Interfaces
     public interface IBondsBase : IObligationsBase
     {
 
-        int? PercentCoupon { get; set; }
+        decimal? PercentCoupon { get; set; }
 
         decimal? PercentYTM { get; set; }
